Support whereLike and whereILike
