+++ lib/dialects/mysql/query/mysql-querycompiler.js
@@ -153,6 +153,20 @@ class QueryCompiler_MySQL extends QueryCompiler {
           );
     return `limit ${limit}`;
   }
+
+  whereLike(statement) {
+    return `${this._columnClause(statement)} ${this._not(
+      statement,
+      'like '
+    )}${this._valueClause(statement)} COLLATE utf8_bin`;
+  }
+
+  whereILike(statement) {
+    return `${this._columnClause(statement)} ${this._not(
+      statement,
+      'ilike '
+    )}${this._valueClause(statement)}`;
+  }
 }
 
 // Set the QueryBuilder & QueryCompiler on the client object,
