+++ lib/query/method-constants.js
@@ -25,6 +25,8 @@ module.exports = [
   'fullOuterJoin',
   'crossJoin',
   'where',
+  'whereLike',
+  'whereILike',
   'andWhere',
   'orWhere',
   'whereNot',
