+++ lib/query/querybuilder.js
@@ -611,6 +611,28 @@ class Builder extends EventEmitter {
     return this._bool('or').whereNotBetween(column, values);
   }
 
+  _whereLike(type, column, values) {
+    this._statements.push({
+      grouping: 'where',
+      type: type,
+      column,
+      value: values,
+      not: this._not(),
+      bool: this._bool(),
+    });
+    return this;
+  }
+
+  // Adds a `where like` clause to the query.
+  whereLike(column, values) {
+    return this._whereLike('whereLike', column, values);
+  }
+
+  // Adds a `where ilike` clause to the query.
+  whereILike(column, values) {
+    return this._whereLike('whereILike', column, values);
+  }
+
   // Adds a `group by` clause to the query.
   groupBy(item) {
     if (item && item.isRawInstance) {
