+++ lib/query/querycompiler.js
@@ -882,7 +882,23 @@ class QueryCompiler {
   // Where Clause
   // ------
 
-  whereIn(statement) {
+  _valueClause(statement) {
+    return statement.asColumn
+      ? wrap_(
+          statement.value,
+          undefined,
+          this.builder,
+          this.client,
+          this.bindingsHolder
+        )
+      : this.client.parameter(
+          statement.value,
+          this.builder,
+          this.bindingsHolder
+        );
+  }
+
+  _columnClause(statement) {
     let columns;
     if (Array.isArray(statement.column)) {
       columns = `(${columnize_(
@@ -900,13 +916,33 @@ class QueryCompiler {
         this.bindingsHolder
       );
     }
+    return columns;
+  }
 
+  whereIn(statement) {
     const values = this.client.values(
       statement.value,
       this.builder,
       this.bindingsHolder
     );
-    return `${columns} ${this._not(statement, 'in ')}${values}`;
+    return `${this._columnClause(statement)} ${this._not(
+      statement,
+      'in '
+    )}${values}`;
+  }
+
+  whereLike(statement) {
+    return `${this._columnClause(statement)} ${this._not(
+      statement,
+      'like '
+    )}${this._valueClause(statement)}`;
+  }
+
+  whereILike(statement) {
+    return `${this._columnClause(statement)} ${this._not(
+      statement,
+      'ilike '
+    )}${this._valueClause(statement)}`;
   }
 
   whereNull(statement) {
@@ -942,19 +978,7 @@ class QueryCompiler {
         this.bindingsHolder
       ) +
       ' ' +
-      (statement.asColumn
-        ? wrap_(
-            statement.value,
-            undefined,
-            this.builder,
-            this.client,
-            this.bindingsHolder
-          )
-        : this.client.parameter(
-            statement.value,
-            this.builder,
-            this.bindingsHolder
-          ))
+      this._valueClause(statement)
     );
   }
