+++ test/integration2/query/select/selects.spec.js
@@ -5,7 +5,6 @@ const { expect } = require('chai');
 const assert = require('assert');
 const Runner = require('../../../../lib/execution/runner');
 
-const { TEST_TIMESTAMP } = require('../../../util/constants');
 const {
   isMysql,
   isPostgreSQL,
@@ -681,457 +680,6 @@ describe('Selects', function () {
         await knex.schema.dropTable('OrderByNullTest');
       });
 
-      describe('simple "where" cases', function () {
-        it('allows key, value', function () {
-          return knex('accounts')
-            .where('id', 1)
-            .select('first_name', 'last_name')
-            .testSql(function (tester) {
-              tester(
-                'mysql',
-                'select `first_name`, `last_name` from `accounts` where `id` = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-              tester(
-                'pg',
-                'select "first_name", "last_name" from "accounts" where "id" = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-              tester(
-                'pgnative',
-                'select "first_name", "last_name" from "accounts" where "id" = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-              tester(
-                'pg-redshift',
-                'select "first_name", "last_name" from "accounts" where "id" = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-              tester(
-                'sqlite3',
-                'select `first_name`, `last_name` from `accounts` where `id` = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-              tester(
-                'oracledb',
-                'select "first_name", "last_name" from "accounts" where "id" = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-              tester(
-                'mssql',
-                'select [first_name], [last_name] from [accounts] where [id] = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-            });
-        });
-
-        it('allows key, operator, value', function () {
-          return knex('accounts')
-            .where('id', 1)
-            .select('first_name', 'last_name')
-            .testSql(function (tester) {
-              tester(
-                'mysql',
-                'select `first_name`, `last_name` from `accounts` where `id` = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-              tester(
-                'pg',
-                'select "first_name", "last_name" from "accounts" where "id" = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-              tester(
-                'pgnative',
-                'select "first_name", "last_name" from "accounts" where "id" = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-              tester(
-                'pg-redshift',
-                'select "first_name", "last_name" from "accounts" where "id" = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-              tester(
-                'sqlite3',
-                'select `first_name`, `last_name` from `accounts` where `id` = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-              tester(
-                'oracledb',
-                'select "first_name", "last_name" from "accounts" where "id" = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-              tester(
-                'mssql',
-                'select [first_name], [last_name] from [accounts] where [id] = ?',
-                [1],
-                [
-                  {
-                    first_name: 'Test',
-                    last_name: 'User',
-                  },
-                ]
-              );
-            });
-        });
-
-        it('allows selecting columns with an array', function () {
-          return knex('accounts')
-            .where('id', '>', 1)
-            .select(['email', 'logins'])
-            .testSql(function (tester) {
-              tester(
-                'mysql',
-                'select `email`, `logins` from `accounts` where `id` > ?',
-                [1]
-              );
-              tester(
-                'pg',
-                'select "email", "logins" from "accounts" where "id" > ?',
-                [1]
-              );
-              tester(
-                'pgnative',
-                'select "email", "logins" from "accounts" where "id" > ?',
-                [1]
-              );
-              tester(
-                'pg-redshift',
-                'select "email", "logins" from "accounts" where "id" > ?',
-                [1]
-              );
-              tester(
-                'sqlite3',
-                'select `email`, `logins` from `accounts` where `id` > ?',
-                [1]
-              );
-              tester(
-                'oracledb',
-                'select "email", "logins" from "accounts" where "id" > ?',
-                [1]
-              );
-              tester(
-                'mssql',
-                'select [email], [logins] from [accounts] where [id] > ?',
-                [1]
-              );
-            });
-        });
-
-        it('allows a hash of where attrs', function () {
-          return knex('accounts')
-            .where({ id: 1 })
-            .select('*')
-            .testSql(function (tester) {
-              tester(
-                'mysql',
-                'select * from `accounts` where `id` = ?',
-                [1],
-                [
-                  {
-                    id: 1,
-                    first_name: 'Test',
-                    last_name: 'User',
-                    email: 'test1@example.com',
-                    logins: 1,
-                    balance: 0,
-                    about: 'Lorem ipsum Dolore labore incididunt enim.',
-                    created_at: TEST_TIMESTAMP,
-                    updated_at: TEST_TIMESTAMP,
-                    phone: null,
-                  },
-                ]
-              );
-              tester(
-                'pg',
-                'select * from "accounts" where "id" = ?',
-                [1],
-                [
-                  {
-                    id: '1',
-                    first_name: 'Test',
-                    last_name: 'User',
-                    email: 'test1@example.com',
-                    logins: 1,
-                    balance: 0,
-                    about: 'Lorem ipsum Dolore labore incididunt enim.',
-                    created_at: TEST_TIMESTAMP,
-                    updated_at: TEST_TIMESTAMP,
-                    phone: null,
-                  },
-                ]
-              );
-              tester(
-                'pgnative',
-                'select * from "accounts" where "id" = ?',
-                [1],
-                [
-                  {
-                    id: '1',
-                    first_name: 'Test',
-                    last_name: 'User',
-                    email: 'test1@example.com',
-                    logins: 1,
-                    balance: 0,
-                    about: 'Lorem ipsum Dolore labore incididunt enim.',
-                    created_at: TEST_TIMESTAMP,
-                    updated_at: TEST_TIMESTAMP,
-                    phone: null,
-                  },
-                ]
-              );
-              tester(
-                'pg-redshift',
-                'select * from "accounts" where "id" = ?',
-                [1],
-                [
-                  {
-                    id: '1',
-                    first_name: 'Test',
-                    last_name: 'User',
-                    email: 'test1@example.com',
-                    logins: 1,
-                    balance: 0,
-                    about: 'Lorem ipsum Dolore labore incididunt enim.',
-                    created_at: TEST_TIMESTAMP,
-                    updated_at: TEST_TIMESTAMP,
-                    phone: null,
-                  },
-                ]
-              );
-              tester(
-                'sqlite3',
-                'select * from `accounts` where `id` = ?',
-                [1],
-                [
-                  {
-                    id: 1,
-                    first_name: 'Test',
-                    last_name: 'User',
-                    email: 'test1@example.com',
-                    logins: 1,
-                    balance: 0,
-                    about: 'Lorem ipsum Dolore labore incididunt enim.',
-                    created_at: TEST_TIMESTAMP,
-                    updated_at: TEST_TIMESTAMP,
-                    phone: null,
-                  },
-                ]
-              );
-              tester(
-                'oracledb',
-                'select * from "accounts" where "id" = ?',
-                [1],
-                [
-                  {
-                    id: 1,
-                    first_name: 'Test',
-                    last_name: 'User',
-                    email: 'test1@example.com',
-                    logins: 1,
-                    balance: 0,
-                    about: 'Lorem ipsum Dolore labore incididunt enim.',
-                    created_at: TEST_TIMESTAMP,
-                    updated_at: TEST_TIMESTAMP,
-                    phone: null,
-                  },
-                ]
-              );
-              tester(
-                'mssql',
-                'select * from [accounts] where [id] = ?',
-                [1],
-                [
-                  {
-                    id: '1',
-                    first_name: 'Test',
-                    last_name: 'User',
-                    email: 'test1@example.com',
-                    logins: 1,
-                    balance: 0,
-                    about: 'Lorem ipsum Dolore labore incididunt enim.',
-                    created_at: TEST_TIMESTAMP,
-                    updated_at: TEST_TIMESTAMP,
-                    phone: null,
-                  },
-                ]
-              );
-            });
-        });
-
-        it('allows where id: undefined or id: null as a where null clause', function () {
-          return knex('accounts')
-            .where({ id: null })
-            .select('first_name', 'email')
-            .testSql(function (tester) {
-              tester(
-                'mysql',
-                'select `first_name`, `email` from `accounts` where `id` is null',
-                [],
-                []
-              );
-              tester(
-                'pg',
-                'select "first_name", "email" from "accounts" where "id" is null',
-                [],
-                []
-              );
-              tester(
-                'pgnative',
-                'select "first_name", "email" from "accounts" where "id" is null',
-                [],
-                []
-              );
-              tester(
-                'pg-redshift',
-                'select "first_name", "email" from "accounts" where "id" is null',
-                [],
-                []
-              );
-              tester(
-                'sqlite3',
-                'select `first_name`, `email` from `accounts` where `id` is null',
-                [],
-                []
-              );
-              tester(
-                'oracledb',
-                'select "first_name", "email" from "accounts" where "id" is null',
-                [],
-                []
-              );
-              tester(
-                'mssql',
-                'select [first_name], [email] from [accounts] where [id] is null',
-                [],
-                []
-              );
-            });
-        });
-
-        it('allows where id = 0', function () {
-          return knex('accounts')
-            .where({ id: 0 })
-            .select()
-            .testSql(function (tester) {
-              tester(
-                'mysql',
-                'select * from `accounts` where `id` = ?',
-                [0],
-                []
-              );
-              tester('pg', 'select * from "accounts" where "id" = ?', [0], []);
-              tester(
-                'pgnative',
-                'select * from "accounts" where "id" = ?',
-                [0],
-                []
-              );
-              tester(
-                'pg-redshift',
-                'select * from "accounts" where "id" = ?',
-                [0],
-                []
-              );
-              tester(
-                'sqlite3',
-                'select * from `accounts` where `id` = ?',
-                [0],
-                []
-              );
-              tester(
-                'oracledb',
-                'select * from "accounts" where "id" = ?',
-                [0],
-                []
-              );
-              tester(
-                'mssql',
-                'select * from [accounts] where [id] = ?',
-                [0],
-                []
-              );
-            });
-        });
-      });
-
       it('#1276 - Dates NULL should be returned as NULL, not as new Date(null)', function () {
         return knex.schema
           .dropTableIfExists('DatesTest')
@@ -1257,328 +805,6 @@ describe('Selects', function () {
         });
       });
 
-      it('does "orWhere" cases', function () {
-        return knex('accounts')
-          .where('id', 1)
-          .orWhere('id', '>', 2)
-          .select('first_name', 'last_name');
-      });
-
-      it('does "andWhere" cases', function () {
-        return knex('accounts')
-          .select('first_name', 'last_name', 'about')
-          .where('id', 1)
-          .andWhere('email', 'test1@example.com');
-      });
-
-      it('takes a function to wrap nested where statements', function () {
-        return Promise.all([
-          knex('accounts')
-            .where(function () {
-              this.where('id', 2);
-              this.orWhere('id', 3);
-            })
-            .select('*'),
-        ]);
-      });
-
-      it('handles "where in" cases', function () {
-        return Promise.all([
-          knex('accounts').whereIn('id', [1, 2, 3]).select(),
-        ]);
-      });
-
-      it('handles "or where in" cases', function () {
-        return knex('accounts')
-          .where('email', 'test1@example.com')
-          .orWhereIn('id', [2, 3, 4])
-          .select();
-      });
-
-      it('handles multi-column "where in" cases', async function () {
-        await knex('composite_key_test').insert([
-          {
-            column_a: 1,
-            column_b: 1,
-            details: 'One, One, One',
-            status: 1,
-          },
-          {
-            column_a: 1,
-            column_b: 2,
-            details: 'One, Two, Zero',
-            status: 0,
-          },
-          {
-            column_a: 2,
-            column_b: 2,
-            details: 'Two, Two, Zero',
-            status: 0,
-          },
-        ]);
-
-        if (!isMssql(knex)) {
-          await knex('composite_key_test')
-            .whereIn(
-              ['column_a', 'column_b'],
-              [
-                [1, 1],
-                [1, 2],
-              ]
-            )
-            .orderBy('status', 'desc')
-            .select()
-            .testSql(function (tester) {
-              tester(
-                'mysql',
-                'select * from `composite_key_test` where (`column_a`, `column_b`) in ((?, ?), (?, ?)) order by `status` desc',
-                [1, 1, 1, 2],
-                [
-                  {
-                    column_a: 1,
-                    column_b: 1,
-                    details: 'One, One, One',
-                    status: 1,
-                  },
-                  {
-                    column_a: 1,
-                    column_b: 2,
-                    details: 'One, Two, Zero',
-                    status: 0,
-                  },
-                ]
-              );
-              tester(
-                'pg',
-                'select * from "composite_key_test" where ("column_a", "column_b") in ((?, ?), (?, ?)) order by "status" desc',
-                [1, 1, 1, 2],
-                [
-                  {
-                    column_a: 1,
-                    column_b: 1,
-                    details: 'One, One, One',
-                    status: 1,
-                  },
-                  {
-                    column_a: 1,
-                    column_b: 2,
-                    details: 'One, Two, Zero',
-                    status: 0,
-                  },
-                ]
-              );
-              tester(
-                'pgnative',
-                'select * from "composite_key_test" where ("column_a", "column_b") in ((?, ?), (?, ?)) order by "status" desc',
-                [1, 1, 1, 2],
-                [
-                  {
-                    column_a: 1,
-                    column_b: 1,
-                    details: 'One, One, One',
-                    status: 1,
-                  },
-                  {
-                    column_a: 1,
-                    column_b: 2,
-                    details: 'One, Two, Zero',
-                    status: 0,
-                  },
-                ]
-              );
-              tester(
-                'pg-redshift',
-                'select * from "composite_key_test" where ("column_a", "column_b") in ((?, ?), (?, ?)) order by "status" desc',
-                [1, 1, 1, 2],
-                [
-                  {
-                    column_a: 1,
-                    column_b: 1,
-                    details: 'One, One, One',
-                    status: 1,
-                  },
-                  {
-                    column_a: 1,
-                    column_b: 2,
-                    details: 'One, Two, Zero',
-                    status: 0,
-                  },
-                ]
-              );
-              tester(
-                'oracledb',
-                'select * from "composite_key_test" where ("column_a", "column_b") in ((?, ?), (?, ?)) order by "status" desc',
-                [1, 1, 1, 2],
-                [
-                  {
-                    column_a: 1,
-                    column_b: 1,
-                    details: 'One, One, One',
-                    status: 1,
-                  },
-                  {
-                    column_a: 1,
-                    column_b: 2,
-                    details: 'One, Two, Zero',
-                    status: 0,
-                  },
-                ]
-              );
-              tester(
-                'sqlite3',
-                'select * from `composite_key_test` where (`column_a`, `column_b`) in ( values (?, ?), (?, ?)) order by `status` desc',
-                [1, 1, 1, 2],
-                [
-                  {
-                    column_a: 1,
-                    column_b: 1,
-                    details: 'One, One, One',
-                    status: 1,
-                  },
-                  {
-                    column_a: 1,
-                    column_b: 2,
-                    details: 'One, Two, Zero',
-                    status: 0,
-                  },
-                ]
-              );
-            });
-        }
-      });
-
-      it('handles multi-column "where in" cases with where', function () {
-        if (!isSQLite(knex) && !isMssql(knex)) {
-          return knex('composite_key_test')
-            .where('status', 1)
-            .whereIn(
-              ['column_a', 'column_b'],
-              [
-                [1, 1],
-                [1, 2],
-              ]
-            )
-            .select()
-            .testSql(function (tester) {
-              tester(
-                'mysql',
-                'select * from `composite_key_test` where `status` = ? and (`column_a`, `column_b`) in ((?, ?), (?, ?))',
-                [1, 1, 1, 1, 2],
-                [
-                  {
-                    column_a: 1,
-                    column_b: 1,
-                    details: 'One, One, One',
-                    status: 1,
-                  },
-                ]
-              );
-              tester(
-                'pg',
-                'select * from "composite_key_test" where "status" = ? and ("column_a", "column_b") in ((?, ?), (?, ?))',
-                [1, 1, 1, 1, 2],
-                [
-                  {
-                    column_a: 1,
-                    column_b: 1,
-                    details: 'One, One, One',
-                    status: 1,
-                  },
-                ]
-              );
-              tester(
-                'pgnative',
-                'select * from "composite_key_test" where "status" = ? and ("column_a", "column_b") in ((?, ?), (?, ?))',
-                [1, 1, 1, 1, 2],
-                [
-                  {
-                    column_a: 1,
-                    column_b: 1,
-                    details: 'One, One, One',
-                    status: 1,
-                  },
-                ]
-              );
-              tester(
-                'pg-redshift',
-                'select * from "composite_key_test" where "status" = ? and ("column_a", "column_b") in ((?, ?), (?, ?))',
-                [1, 1, 1, 1, 2],
-                [
-                  {
-                    column_a: 1,
-                    column_b: 1,
-                    details: 'One, One, One',
-                    status: 1,
-                  },
-                ]
-              );
-              tester(
-                'oracledb',
-                'select * from "composite_key_test" where "status" = ? and ("column_a", "column_b") in ((?, ?), (?, ?))',
-                [1, 1, 1, 1, 2],
-                [
-                  {
-                    column_a: 1,
-                    column_b: 1,
-                    details: 'One, One, One',
-                    status: 1,
-                  },
-                ]
-              );
-            });
-        }
-      });
-
-      it('handles "where exists"', function () {
-        return knex('accounts')
-          .whereExists(function () {
-            this.select('id').from('test_table_two').where({ id: 1 });
-          })
-          .select();
-      });
-
-      it('handles "where between"', function () {
-        return knex('accounts').whereBetween('id', [1, 100]).select();
-      });
-
-      it('handles "or where between"', function () {
-        return knex('accounts')
-          .whereBetween('id', [1, 100])
-          .orWhereBetween('id', [200, 300])
-          .select();
-      });
-
-      it('does where(raw)', function () {
-        if (isOracle(knex)) {
-          // special case for oracle
-          return knex('accounts')
-            .whereExists(function () {
-              this.select(knex.raw(1))
-                .from('test_table_two')
-                .where(
-                  knex.raw('"test_table_two"."account_id" = "accounts"."id"')
-                );
-            })
-            .select();
-        } else {
-          return knex('accounts')
-            .whereExists(function () {
-              this.select(knex.raw(1))
-                .from('test_table_two')
-                .where(knex.raw('test_table_two.account_id = accounts.id'));
-            })
-            .select();
-        }
-      });
-
-      it('does sub-selects', function () {
-        return knex('accounts')
-          .whereIn('id', function () {
-            this.select('account_id').from('test_table_two').where('status', 1);
-          })
-          .select('first_name', 'last_name');
-      });
-
       describe('recursive CTE support', function () {
         before(async function () {
           await knex.schema.dropTableIfExists('rcte');
@@ -1625,45 +851,6 @@ describe('Selects', function () {
           expect(names).to.contain('parent');
           expect(names).not.to.contain('nope');
         });
-      });
-
-      it('supports the <> operator', function () {
-        return knex('accounts').where('id', '<>', 2).select('email', 'logins');
-      });
-
-      it('Allows for knex.Raw passed to the `where` clause', function () {
-        if (isOracle(knex)) {
-          return knex('accounts')
-            .where(knex.raw('"id" = 2'))
-            .select('email', 'logins');
-        } else {
-          return knex('accounts')
-            .where(knex.raw('id = 2'))
-            .select('email', 'logins');
-        }
-      });
-
-      it('Retains array bindings, #228', function () {
-        const raw = knex.raw(
-          'select * from table t where t.id = ANY( ?::int[] )',
-          [[1, 2, 3]]
-        );
-        const raw2 = knex.raw('select "stored_procedure"(?, ?, ?)', [
-          1,
-          2,
-          ['a', 'b', 'c'],
-        ]);
-        const expected1 = [[1, 2, 3]];
-        const expected2 = [1, 2, ['a', 'b', 'c']];
-        expect(raw.toSQL().bindings).to.eql(
-          knex.client.prepBindings(expected1)
-        );
-        expect(raw2.toSQL().bindings).to.eql(
-          knex.client.prepBindings(expected2)
-        );
-        //Also expect raw's bindings to not have been modified by calling .toSQL() (preserving original bindings)
-        expect(raw.bindings).to.eql(expected1);
-        expect(raw2.bindings).to.eql(expected2);
       });
 
       it('always returns the response object from raw', function () {
